Add user jwt login, Add DAO classes

+++ src/main/java/com/devgd/melonclone/domain/album/domain/AlbumEntity.java
@@ -22,30 +22,30 @@ import lombok.Getter;
 public class AlbumEntity extends BaseEntity implements Serializable {
 	@Id
 	@GeneratedValue(strategy= GenerationType.IDENTITY)
-	private Long album_id;
+	private Integer album_id;
 
 	@Column(nullable = false)
-	private Long album_artist_id;
+	private Integer album_artist_id;
 
 	@Column(length = 45, nullable = false)
 	private String album_name;
 
 	@Column(nullable = false)
-	private Long album_like;
+	private Integer album_like;
 
 	@Column(length = 255, nullable = false)
 	private String album_jacket;
 
 	@Column(nullable = false)
-	private Long album_category;
+	private Integer album_category;
 
 	@Column(nullable = false)
 	private LocalDateTime create_date;
 
 	@Builder
-	public AlbumEntity(Long album_id, Long album_artist_id, 
-			String album_name, Long album_like,
-			String album_jacket, Long album_category, LocalDateTime create_date) {
+	public AlbumEntity(Integer album_id, Integer album_artist_id, 
+			String album_name, Integer album_like,
+			String album_jacket, Integer album_category, LocalDateTime create_date) {
 		this.album_id = album_id;
 		this.album_artist_id = album_artist_id;
 		this.album_name = album_name;
