+++ src/main/java/com/devgd/melonclone/domain/music/domain/MusicId.java
@@ -14,19 +14,19 @@ public class MusicId implements Serializable {
 	private static final long serialVersionUID = -390950220271593086L;
 
 	@Column(nullable = false)
-	private final Long music_id;
+	private final Integer music_id;
 
 	@Column(nullable = false)
-	private final Long music_artist_id;
+	private final Integer music_artist_id;
 
 	@Column(nullable = false)
-	private final Long music_album_id;
+	private final Integer music_album_id;
 
 	@Column(nullable = false)
-	private final Long music_category_id;
+	private final Integer music_category_id;
 
 	@Builder
-	public MusicId(Long music_id, Long music_artist_id, Long music_album_id, Long music_category_id) {
+	public MusicId(Integer music_id, Integer music_artist_id, Integer music_album_id, Integer music_category_id) {
 		this.music_id = music_id;
 		this.music_artist_id = music_artist_id;
 		this.music_album_id = music_album_id;
